+++ NEWS.md
@@ -6,6 +6,13 @@
 - Classes `SimulationBatchRunValues` and `SimulationBatchOptions` are not exported any more.
 They should not be used directly.
 
+## Major changes
+- `DataImporterConfiguration` now supports getting and setting of the Lower Limit 
+of Quantification (LLOQ) column name via
+a new field `lloqColumn`. If the column name is not set (value `NULL`), LLOQ values
+will be imported from the measurement column if values are written in the form '< xxx' (e.g., '<0.001').
+Otherwise, the values will be imported from the specified column (\#722)
+
 ## Minor improvements and bug fixes
 
 - Improved print outputs for all classes